==> aeroscope/01_wikipedia_parser/03-routes_processing.ipynb <==
# Wikipedia route database construction

This notebook uses the route list retrieved by 02-routes_parsing.ipynb  

- the destination url is replaced by the canonical url to allow for a clean merge with airport data 
- the airline url is used to find informations on airlines


import numpy as np
import requests
from requests.adapters import HTTPAdapter, Retry
from bs4 import BeautifulSoup
from string import Template
import pandas as pd
from tqdm.notebook import tqdm

tqdm.pandas()

## Step 1 : find the destination airports

**Fisrt run**:  
Indeed, the wikipedia list of airport is not exhaustive and some airports are misisng. Hence, we try a first merge of the routes with the airports.  
The missing airports links (only for big airports (k> 14 routes), otherwise the iterative process could go on for a long time) are reinjected in the 01-airport_parsing notebook, added to the airports database, and to 02-routes_parsing to find their routes. 

**Other runs**:
The resulting new list is processed again in this notebook, there are still "missing airports" but those should not be big airports, wich is fine for our application.

raw_routes = pd.read_csv("data/wikipedia_relations_new24_04.csv")
raw_routes.drop(columns=["Unnamed: 0"], index=1, inplace=True)
raw_routes.columns = ["airline", "origin", "destination", "type"]

One problem that emerged during this preparation is that wiki can have several different url at once. However, a unique 'canonical' url is always present. 
We need to associate the canonical url of each airport. 

 
Process: deduplicate airports links, open page and get canonical url.  
NB: the origin of each route is already the cannonical url, since the canonical url (instead of the url) of each airport is catched during airport html parsing. 


airports_nodups = pd.DataFrame(raw_routes["destination"].drop_duplicates())

def find_canonical_url(airport_link):
    headers = {"User-Agent": "AirProjectBot/0.0 (antoine732@hotmail.fr)"}
    # response = requests.get(airline_link, headers=headers)

    try:
        session = requests.Session()
        retries = Retry(
            total=5,
            backoff_factor=1,
            status_forcelist=[429],
            respect_retry_after_header=True,
        )
        session.mount("https://", HTTPAdapter(max_retries=retries))
        response = session.get(airport_link, headers=headers)
    except:
        response = np.nan

    if response is not np.nan:
        soup = BeautifulSoup(response.content, "html.parser")
        wdpa_link = soup.find("link", {"rel": "canonical"}).get("href")

    else:
        wdpa_link = airport_link  # we shouldn't get there
        print("No canonical url found!")
    return wdpa_link

airport_canonical = airports_nodups.progress_apply(
    lambda x: find_canonical_url(x["destination"]), axis=1, result_type="expand"
)

airport_canonical

airports_nodups["canonical_destination"] = airport_canonical

# airports_nodups.to_csv('data/airport_canonical_24_04.csv')
airports_nodups = pd.read_csv("data/airport_canonical_24_04.csv")

raw_routes = raw_routes.merge(
    airports_nodups, left_on="destination", right_on="destination", how="left"
)

raw_routes

Now, we merge the routes dataframe with the airport data from 01 notebook.
This is JUST to look for missing airports: top proper merge with aggregated airport data is done afterwards (02_airport_features)

As metionned in the introduction, this notebook is ran TWICE:

- First run --> only airports retrieved from wikipedia continental list of airports. Airports are merged with routes.  
Sometimes the destination link is not in the airport database (origin is always, by construction).  
It is stored in a missing_airport files in the following, that is processed in 1 and 02 notebooks.
Unknown airports are added only if they have more than 14 routes. 

- After the first run, missing airports should only be "minor" airports, with less than 14 destinations. Nevertheless, they are processed in 01 (not in 02) once again if they have more than 2 routes (1 routers were not included because it includes too specific wikipedia pages (natural reserves, oil rigs, ...) not adapted for the querry. It means we still look for these destinations features with wikipedia. Their routes are not parsed however, but this could be done. *NB: it could be done, but this second step was added as a refinement later in the process, and doing so would erase include two route parsing dates in the database, which shoud be avoided. It only removes routes between TWO airports of this category, which should be a very marginal proportion of the database and would add much noise.*


nam_airport_df = pd.read_csv("data/n_amer_arpt_v2.csv")
sam_airport_df = pd.read_csv("data/s_amer_arpt_v2.csv")
eu_airport_df = pd.read_csv("data/eu_arpt_v2.csv")
af_airport_df = pd.read_csv("data/af_arpt_v2.csv")
as_airport_df = pd.read_csv("data/as_arpt_v2.csv")
oc_airport_df = pd.read_csv("data/oc_arpt_v2.csv")
special_airport_df = pd.read_csv("data/pb_arpt_v2.csv")

# Only after first run
missing_airport_df = pd.read_csv("data/missing_arpt_v2.csv")

# Only after second run

missing_airport_extra_df = pd.read_csv("data/missing_arpt_extra.csv")
# conbine airports in a single dataframe.

all_airport_df = (
    pd.concat(
        [
            nam_airport_df,
            sam_airport_df,
            eu_airport_df,
            af_airport_df,
            as_airport_df,
            oc_airport_df,
            special_airport_df,
            missing_airport_df,
            missing_airport_extra_df,
        ],
        axis=0,
    )
    .reset_index()
    .drop(columns=["index", "Unnamed: 0"], axis=1)
)

reduced_airport_df = all_airport_df[
    [
        "icao",
        "iata",
        "max_population",
        "max_passengers19",
        "maxpax",
        "wdpa_link",
        "wpda_iata",
        "wpda_icao",
    ]
]

reduced_airport_df.columns = [
    "O_icao",
    "O_iata",
    "O_max_population",
    "O_max_passengers19",
    "O_maxpax",
    "O_wdpa_link",
    "O_wpda_iata",
    "O_wpda_icao",
]
routes_w_departures = raw_routes.merge(
    reduced_airport_df, left_on="origin", right_on="O_wdpa_link", how="left"
)

reduced_airport_df.columns = [
    "D_icao",
    "D_iata",
    "D_max_population",
    "D_max_passengers19",
    "D_maxpax",
    "D_wdpa_link",
    "D_wpda_iata",
    "D_wpda_icao",
]
routes_full = routes_w_departures.merge(
    reduced_airport_df,
    left_on="canonical_destination",
    right_on="D_wdpa_link",
    how="left",
)

routes_full.dropna(subset="O_iata", inplace=True)

missing_airports = (
    routes_full[routes_full["D_wdpa_link"].isna()]
    .groupby(["destination"])["airline"]
    .count()
    .sort_values()
)

# missing_airports.index
missing_airports = missing_airports.reset_index()
missing_airports = missing_airports[missing_airports["airline"] > 3]

# storing the additional misisng airports in a csv => routes won't be searched frm these airports, but their information in parsed to improve data quality and remove unknwon airports in the database
# missing_airports.reset_index()['destination'].to_csv('data/extra_airport_refs.csv')

routes_full = routes_full[~routes_full["D_wdpa_link"].isna()][
    ["airline", "origin", "canonical_destination", "type"]
]
routes_full

## Step 2: add airlines codes


Let's list the unique airlines links before queryng their IATA ID 

airline_nodups = routes_full["airline"].drop_duplicates()

def find_airlines_iata(airline_link):

    headers = {"User-Agent": "AirProjectBot/0.0 (antoine732@hotmail.fr)"}
    # response = requests.get(airline_link, headers=headers)

    try:
        session = requests.Session()
        retries = Retry(
            total=5,
            backoff_factor=1,
            status_forcelist=[429],
            respect_retry_after_header=True,
        )
        session.mount("https://", HTTPAdapter(max_retries=retries))
        response = session.get(airline_link, headers=headers)
    except:
        response = np.nan

    airline_name = np.nan
    wpda_iata = np.nan
    wpda_icao = np.nan

    if response is not np.nan:
        soup = BeautifulSoup(response.content, "html.parser")

        # Find the airport information card and name (page title)
        vcard = soup.find("table", class_="infobox vcard")
        title = soup.find(class_="mw-page-title-main")
        if title is not None:
            airline_name = title.text

        if vcard is not None:
            # Extract interesting information from the vcard when available
            # Extract the "nickname"; it is html name of the iata and icao field of Wiki pages

            if vcard.find("td", class_="infobox-full-data") is not None:
                table = vcard.find("td", class_="infobox-full-data").find("table")

                if table is not None:
                    row = []
                    for tr in table.find_all("tr"):
                        for td in tr.find_all("td"):
                            row.append(td.text.strip("\n"))

                wpda_iata = row[0]
                wpda_icao = row[1]

    return [airline_link, airline_name, wpda_iata, wpda_icao]

pd.DataFrame(airline_nodups)

airline_data = pd.DataFrame(airline_nodups).progress_apply(
    lambda x: find_airlines_iata(x["airline"]), axis=1, result_type="expand"
)

airline_data.reset_index(inplace=True)

airline_data.drop(columns="index", axis=1, inplace=True)

airline_data.columns = ["airline_link", "airline_name", "airline_iata", "airline_icao"]
airline_data

airline_data[airline_data["airline_icao"].isna()]

## Manual correction of small problemns encountered before ==> not worth spending time on a robust code for 5 wrong airlines

airline_data.iloc[5]["airline_name"] = "JetBlue"
airline_data.iloc[365]["airline_name"] = "EasyJet"
airline_data.iloc[232]["airline_name"] = "iAero Airways"
airline_data.iloc[372]["airline_name"] = "airBaltic"
airline_data.iloc[568]["airline_name"] = "flyadeal"
airline_data.iloc[607]["airline_name"] = "airBaltic"
airline_data.iloc[628][
    "airline_link"
] = "https://en.wikipedia.orghttps://fr.wikipedia.org/wiki/Amelia_International"
airline_data.iloc[628]["airline_name"] = "Amelia International"
airline_data.iloc[628]["airline_iata"] = "NL"
airline_data.iloc[628]["airline_icao"] = "AEH"

## The following airlines exists but they are aggregates of operational divisions, each being a proper airline.
## Wiki doesn't focus on that and saddly it isn't possible to know which division fly each route.
## Flights are affected to their parent airlines.


# airline_data.iloc[364]['airline_iata']='U2'
# airline_data.iloc[364]['airline_icao']='EZY'

airline_data.iloc[41]["airline_iata"] = "AA"
airline_data.iloc[41]["airline_icao"] = "AAL"

airline_data.iloc[22]["airline_iata"] = "UA"
airline_data.iloc[22]["airline_icao"] = "UAL"

airline_data.iloc[14]["airline_iata"] = "DL"
airline_data.iloc[14]["airline_icao"] = "DAL"

airline_data.iloc[9]["airline_iata"] = "AC"
airline_data.iloc[9]["airline_icao"] = "ACA"

# airline_data.to_csv('data/airline_data_24_04.csv')
# airline_data=pd.read_csv('data/airline_data_24_04.csv')

Now , the airport database constructed is used to convert the route O&D airports wikipedia links into a global df with relevant informations on O&D airports

routes_w_airlines = routes_full.merge(
    airline_data, left_on="airline", right_on="airline_link", how="left"
)

routes_w_airlines[routes_w_airlines["airline_iata"].isna()].groupby("airline_name")[
    "airline_link"
].count()

## Safety check, no airline with a significant number of routes should have NaN iata.
## I consider airlines with less than 30 routes not worth the investigation (the dataset contains 100k routes).

routes_w_airlines

routes_w_airlines = routes_w_airlines.drop(columns={"airline_link"})
routes_w_airlines["airline_iata"] = (
    routes_w_airlines["airline_iata"]
    .str.split("[")
    .str[0]
    .str.split(";")
    .str[0]
    .str.split("*")
    .str[0]
)
routes_w_airlines["airline_icao"] = (
    routes_w_airlines["airline_icao"]
    .str.split("[")
    .str[0]
    .str.split(";")
    .str[0]
    .str.split("*")
    .str[0]
)
routes_w_airlines.loc[
    routes_w_airlines["airline_iata"] == "See Operators", "airline_iata"
] = "WS"
routes_w_airlines.loc[
    routes_w_airlines["airline_icao"] == "See Operators", "airline_icao"
] = "WJA"

routes_w_airlines.to_csv("data/wiki_route_db_26_09.csv")

routes_w_airlines

The following steps are in 02 Folder
